python_requirements.txt
Package              Version
-------------------- ---------------
aiohappyeyeballs     2.4.0
aiohttp              3.10.5
aiosignal            1.3.1
annotated-types      0.7.0
anyio                4.4.0
attrs                24.2.0
beautifulsoup4       4.12.3
bitarray             3.0.0
certifi              2024.8.30
charset-normalizer   3.3.2
ckzg                 2.0.1
contourpy            1.3.0
cssselect            1.2.0
cycler               0.12.1
cytoolz              1.0.0
deep-translator      1.11.4
Deprecated           1.2.14
eth_abi              5.1.0
eth-account          0.13.4
eth-hash             0.7.0
eth-keyfile          0.8.1
eth-keys             0.6.0
eth-rlp              2.1.0
eth-typing           5.0.1
eth-utils            5.1.0
fonttools            4.53.1
frozenlist           1.4.1
fuzzywuzzy           0.18.0
h11                  0.14.0
hexbytes             1.2.1
httpcore             1.0.5
httpx                0.27.2
idna                 3.8
iniconfig            2.0.0
jellyfish            1.1.0
kiwisolver           1.4.7
Levenshtein          0.26.1
lxml                 5.3.0
matplotlib           3.9.2
mss                  9.0.1
multidict            6.0.5
nodriver             0.36
numpy                2.1.1
packaging            24.1
pandas               2.2.2
parsimonious         0.10.0
pillow               10.4.0
pip                  24.3.1
pluggy               1.5.0
poly_eip712_structs  0.0.1
prettytable          3.11.0
psycopg2-binary      2.9.9
py_clob_client       0.18.0
py_order_utils       0.3.2
pycryptodome         3.21.0
pydantic             2.8.2
pydantic_core        2.20.1
pyparsing            3.1.4
pytest               8.3.3
python-dateutil      2.9.0.post0
python-dotenv        1.0.1
python-Levenshtein   0.26.1
python-telegram-bot  21.5
pytz                 2024.2
RapidFuzz            3.10.1
regex                2024.11.6
requests             2.32.3
rlp                  4.0.1
ruff                 0.6.4
six                  1.16.0
sniffio              1.3.1
soupsieve            2.6
toolz                1.0.0
types-beautifulsoup4 4.12.0.20241020
types-html5lib       1.1.11.20241018
types-lxml           2024.11.8
typing_extensions    4.12.2
tzdata               2024.1
urllib3              2.2.2
wcwidth              0.2.13
websockets           13.0.1
wrapt                1.16.0
yarl                 1.10.0
